practice script

+++ SQL for Data Science - Coursera/table_schema.sql
@@ -1,69 +1,70 @@
 CREATE DATABASE production;
+
 USE production;
 
 CREATE TABLE cheese_production (
-    Year INTEGER,
-    Period TEXT,
-    Geo_Level TEXT,
-    State_ANSI INTEGER,
-    Commodity_ID INTEGER,
-    Domain TEXT,
-    Value INTEGER
+    Year INTEGER,
+    Period TEXT,
+    Geo_Level TEXT,
+    State_ANSI INTEGER,
+    Commodity_ID INTEGER,
+    Domain TEXT,
+    Value BIGINT
 );
 
 
 CREATE TABLE honey_production (
-    Year INTEGER,
-    Geo_Level TEXT,
-    State_ANSI INTEGER,
-    Commodity_ID INTEGER,
-    Value INTEGER
+    Year INTEGER,
+    Geo_Level TEXT,
+    State_ANSI INTEGER,
+    Commodity_ID INTEGER,
+    Value BIGINT
 );
 
 
 CREATE TABLE milk_production (
-    Year INTEGER,
-    Period TEXT,
-    Geo_Level TEXT,
-    State_ANSI INTEGER,
-    Commodity_ID INTEGER,
-    Domain TEXT,
-    Value INTEGER
+    Year INTEGER,
+    Period TEXT,
+    Geo_Level TEXT,
+    State_ANSI INTEGER,
+    Commodity_ID INTEGER,
+    Domain TEXT,
+    Value BIGINT
 );
 
 
 CREATE TABLE coffee_production (
-    Year INTEGER,
-    Period TEXT,
-    Geo_Level TEXT,
-    State_ANSI INTEGER,
-    Commodity_ID INTEGER,
-    Value INTEGER
+    Year INTEGER,
+    Period TEXT,
+    Geo_Level TEXT,
+    State_ANSI INTEGER,
+    Commodity_ID INTEGER,
+    Value BIGINT
 );
 
 
 CREATE TABLE egg_production (
-    Year INTEGER,
-    Period TEXT,
-    Geo_Level TEXT,
-    State_ANSI INTEGER,
-    Commodity_ID INTEGER,
-    Value INTEGER
+    Year INTEGER,
+    Period TEXT,
+    Geo_Level TEXT,
+    State_ANSI INTEGER,
+    Commodity_ID INTEGER,
+    Value BIGINT
 );
 
 
 CREATE TABLE state_lookup (
-    State TEXT,
-    State_ANSI INTEGER
+    State TEXT,
+    State_ANSI INTEGER
 );
 
 
 CREATE TABLE yogurt_production (
-    Year INTEGER,
-    Period TEXT,
-    Geo_Level TEXT,
-    State_ANSI INTEGER,
-    Commodity_ID INTEGER,
-    Domain TEXT,
-    Value INTEGER
+    Year INTEGER,
+    Period TEXT,
+    Geo_Level TEXT,
+    State_ANSI INTEGER,
+    Commodity_ID INTEGER,
+    Domain TEXT,
+    Value BIGINT
 );
